refactor: TaskTableViewCell내 configureCell 로직 viewModel로 이동 #4
- 테스트를 위해 데이터를 가공하는 로직을 viewmodel로 이동
- TaskCell 타입 생성
