+++ ProjectManager/ProjectManager/ViewModel/TaskCell.swift
@@ -1,0 +1,14 @@
+//
+//  TaskCell.swift
+//  ProjectManager
+//
+//  Created by 이차민 on 2022/03/14.
+//
+
+import Foundation
+
+struct TaskCell {
+    var title: String
+    var description: String
+    var deadline: NSAttributedString
+}
